fix: split e2e tests in different files
